New Defragmentation algorithm that discards levels of hold-year
Rather than having the furthest hold coordinate be reserved for the longest
hold, we simply share all the hold coordinates (keeping the rule that the front
(aka 0) is always "drink now"). This solves two problems:

First, it reduces the amount of error in the price coordinate, since duplicate
priced bottles can be placed in the same price row.

Second, it prefers the back slot, reducing the number of bottles that need to be
shifted around when new bottles are added.

Moreover, the visual appeal of balancing these is lost due to the
dimensions. Better to balance the boldness / price coordinates better.